Initial Commit, needs testing and tests still.
