+++ DaemonsMCP.Core/DaemonsMCP.Core.csproj
@@ -1,28 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net9.0-windows7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <AssemblyVersion>1.0.1.1</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="MCPSharp" Version="1.0.11" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.8" />
     <PackageReference Include="PackedTables.Net" Version="1.1.6" />
     <PackageReference Include="System.Formats.Asn1" Version="9.0.8" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\DaemonsMCP.Core\DaemonsMCP.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="daemonsmcp.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
   </ItemGroup>
 
 </Project>
