+++ DaemonsMCP.Core/Models/SecuritySettings.cs
@@ -5,46 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace DaemonsMCP {
-  public class DaemonsMcpConfiguration {
-    [JsonPropertyName("version")]
-    public string Version { get; set; } = "1.0";
-
-    [JsonPropertyName("daemon")]
-    public DaemonSettings Daemon { get; set; } = new();
-
-    [JsonPropertyName("projects")]
-    public List<ProjectConfiguration> Projects { get; set; } = new();
-
-    [JsonPropertyName("security")]
-    public SecuritySettings Security { get; set; } = new();
-  }
-
-  public class DaemonSettings {
-    [JsonPropertyName("name")]
-    public string Name { get; set; } = "DaemonsMCP";
-
-    [JsonPropertyName("version")]
-    public string Version { get; set; } = "1.0.0";
-
-    [JsonPropertyName("logLevel")]
-    public string LogLevel { get; set; } = "Information";
-  }
-
-  public class ProjectConfiguration {
-    [JsonPropertyName("name")]
-    public string Name { get; set; } = "";
-
-    [JsonPropertyName("description")]
-    public string Description { get; set; } = "";
-
-    [JsonPropertyName("path")]
-    public string Path { get; set; } = "";
-
-    [JsonPropertyName("enabled")]
-    public bool Enabled { get; set; } = true;
-  }
-
+namespace DaemonsMCP.Core.Models {
   public class SecuritySettings {
     [JsonPropertyName("allowedExtensions")]
     public List<string> AllowedExtensions { get; set; } = new()
@@ -77,4 +38,5 @@ namespace DaemonsMCP {
     public string MaxFileWriteSize { get; set; } = "5MB";
 
   }
+
 }
