+++ DaemonsMCP.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DaemonsMCP.Tests", "DaemonsMCP.Tests\DaemonsMCP.Tests.csproj", "{5A3CEE25-F03E-471B-82E2-77FB2EF954C4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DaemonsMCP.Core", "DaemonsMCP.Core\DaemonsMCP.Core.csproj", "{C3CA51FE-7D5D-404D-4F1F-72230DE5F972}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{5A3CEE25-F03E-471B-82E2-77FB2EF954C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5A3CEE25-F03E-471B-82E2-77FB2EF954C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5A3CEE25-F03E-471B-82E2-77FB2EF954C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C3CA51FE-7D5D-404D-4F1F-72230DE5F972}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C3CA51FE-7D5D-404D-4F1F-72230DE5F972}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C3CA51FE-7D5D-404D-4F1F-72230DE5F972}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C3CA51FE-7D5D-404D-4F1F-72230DE5F972}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
