--- DaemonsMCP/StringExts.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DaemonsMCP {
-  public static class Sx {
-
-    public static string[] Parse(this string content, string delims) {
-      return content.Split(delims.ToCharArray(), StringSplitOptions.RemoveEmptyEntries);
-    }
-
-    public static string ParseFirst(this string content, string delims) {
-      string[] sr = content.Parse(delims);
-      if (sr.Length > 0) {
-        return sr[0];
-      }
-      return "";
-    }
-    public static string ParseLast(this string content, string delims) {
-      string[] sr = content.Parse(delims);
-      if (sr.Length > 0) {
-        return sr[^1];
-      }
-      return "";
-    }
-
-  }
-}
